requirements.txt
pandas
numpy
scikit-learn
pyyaml
joblib
mlflow
python-dotenv
dvc
dvc-gdrive
Pillow
